Namespace classes and make the project autoload with PSR-4
