+++ composer.json
@@ -18,9 +18,9 @@
 		"php": ">=5.4.0"
 	},
 	"autoload": {
-		"classmap": [
-			"src/"
-		]
+		"psr-4": {
+			"Nbbc\\": "src"
+		}
 	},
 	"autoload-dev": {
 		"psr-4": {
