+++ src/BBCodeLexer.php
